% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibEvalCritO2.R
\name{hsFullSubTitle}
\alias{hsFullSubTitle}
\title{Full Subtitle}
\usage{
hsFullSubTitle(mySub, vecYr)
}
\description{
Given a vector of year numbers a title "of year <year>" 
  or "of years <startYear> to <endYear> is returned
}
\keyword{internal}
