% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibEvalCritO2.R
\name{hsGetYmax}
\alias{hsGetYmax}
\title{Get Ymax}
\usage{
hsGetYmax(matData, myBeside, myExpand = 1.1, dbg = FALSE)
}
\description{
Next highest multiple of appropriate powers of ten, but at least 5
}
\keyword{internal}
