% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibEvalCritO2.R
\name{hsGroupByYearAndKm}
\alias{hsGroupByYearAndKm}
\title{Group By Year And Km}
\usage{
hsGroupByYearAndKm(
  frmData,
  strValField,
  boolDescKm = FALSE,
  boolDescYr = FALSE,
  vecMp = NULL
)
}
\arguments{
\item{frmData}{data.frame with columns "Jahr" and "Spree_km"}

\item{strValField}{Name of value field in input data.frame of which for each Jahr/Spree_km 
group the sum will be calculated.}

\item{boolDescKm}{if TRUE, columns in result matrix will be ordered according to decreasing
"Spree_km" values}

\item{boolDescYr}{if TRUE, columns in result matrix will be ordered according to decreasing
"Jahr" values}

\item{vecMp}{Vector containing pairs of km value and monitoring point name}
}
\value{
Returns the matrix calculated by hsGroupBy2Fields()
}
\description{
Groups input data.frame by its columns "Jahr" and "Spree_km".
}
