% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibEvalCritO2.R
\name{hsPlotAll}
\alias{hsPlotAll}
\title{Plot All}
\usage{
hsPlotAll(
  strDb,
  strTable,
  myScaled = TRUE,
  myReverse = TRUE,
  myBeside = TRUE,
  myCexNames = 1,
  mySub = "",
  dbg = FALSE,
  myWidth = 1,
  myValLabs = FALSE,
  yearsInSub = FALSE,
  lng = "de",
  ...
)
}
\arguments{
\item{strDb}{path to Microsoft Access Database file}

\item{strTable}{name of database table containing the data to plot}

\item{myScaled}{logical. Default: \code{TRUE}}

\item{myReverse}{logical. Default: \code{TRUE}}

\item{myBeside}{if \code{TRUE} (the default), bars are plotted side by side}

\item{myCexNames}{character expansion factor for names}

\item{mySub}{subtitle. Default: ""}

\item{dbg}{if \code{TRUE}, debug messages are shown}

\item{myWidth}{bar width. Default: 1}

\item{myValLabs}{logical. Default: \code{FALSE}}

\item{yearsInSub}{logical. Default: \code{FALSE}}

\item{lng}{language code, one of "en" (English, the default) or "de" (German)}

\item{\dots}{further arguments passed to \code{\link{hsPlotCritEvents}}}
}
\description{
Plot result of evaluation in forms of barplots, with default settings
}
