% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibEvalCritO2.R
\name{hsPlotAllToPdf}
\alias{hsPlotAllToPdf}
\title{Plot All To PDF File}
\usage{
hsPlotAllToPdf(strPdf, ...)
}
\arguments{
\item{strPdf}{path to PDF file to which to plot}

\item{\dots}{further arguments passed to \code{\link{hsPlotAll}}}
}
\description{
Print result of evaluation in forms of barplots into pdf file
}
