% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibEvalCritO2.R
\name{hsPlotCritEvents}
\alias{hsPlotCritEvents}
\title{Plot Result of Evaluation as Barplots}
\usage{
hsPlotCritEvents(
  matData,
  strTitle = "Events",
  mySub = "",
  boolXScaled = FALSE,
  boolDescKm = TRUE,
  boolDescYr = FALSE,
  myBeside = TRUE,
  myWidth = 1,
  myYmax = -1,
  myCexNames = -1,
  myAxis = TRUE,
  myValLabs = FALSE,
  dbg = FALSE,
  yearsInSub = FALSE,
  xlab = "",
  ...
)
}
\description{
Plot Result of Evaluation as Barplots
}
\keyword{internal}
