% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibEvalCritO2.R
\name{hsPlotO2Eval}
\alias{hsPlotO2Eval}
\title{Plot O2 Evaluation}
\usage{
hsPlotO2Eval(dat, main = "Title?", lng = "en")
}
\arguments{
\item{dat}{data frame with columns \emph{Jahr}, \emph{LamEvents}, \emph{2mgEvents},
\emph{LamKalTage}, \emph{2mgKalTage}}

\item{main}{\code{main} plot title}

\item{lng}{language code, one of "en" (English, the default) or "de" (German)}
}
\description{
Plot O2 Evaluation
}
