% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibEvalCritO2.R
\name{hsPlotO2EvalPdf}
\alias{hsPlotO2EvalPdf}
\title{Plot Result of O2 Evaluation to PDF}
\usage{
hsPlotO2EvalPdf(dat, main = "Title?", pdffile = NULL)
}
\arguments{
\item{dat}{data frame with columns \emph{Jahr}, \emph{LamEvents}, \emph{2mgEvents},
\emph{LamKalTage}, \emph{2mgKalTage}}

\item{main}{\code{main} plot title}

\item{pdffile}{path to PDF file to which to plot}
}
\description{
Plot Result of O2 Evaluation to PDF
}
