% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibEvalCritO2.R
\name{hsReformatName}
\alias{hsReformatName}
\title{Reformat Name}
\usage{
hsReformatName(myName)
}
\description{
Label of the form "[<station>: ]km ##.#" is created from \emph{myName} where
  the part in brackets (station name) is optional.
  The part following a semicolon in \emph{myName} is treated as the station 
  name.
}
\keyword{internal}
