% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibEvalCritO2.R
\name{hsReorderMatrix}
\alias{hsReorderMatrix}
\title{Reorder Matrix Columns}
\usage{
hsReorderMatrix(
  matData,
  boolXScaled,
  boolDescKm,
  boolDescYr,
  vecKm,
  vecYr,
  dbg = FALSE
)
}
\description{
Reorder Matrix Columns
}
\keyword{internal}
