% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibEvalCritO2.R
\name{hsTranslate}
\alias{hsTranslate}
\title{Translate}
\usage{
hsTranslate(text.en, lng)
}
\arguments{
\item{text.en}{english text (character vector of length 1 expected)}

\item{lng}{target language: en = English, de = German}
}
\description{
translates \emph{text.en} into target language \emph{lng}.
}
